// sourcecode/BTrak.Packages/snova-hrmangement/module/module-components/src/lib/hrmanagment/models/employee-licence-details-model.ts
export class EmployeeLicenceDetailsModel{
    employeeId: string;
    firstName: string;
    surName: string;
    userName: string;
    email: string;
    licenceTypeId: string;
    licenceType: string;
    licenceNumber: string;
    issuedDate: Date;
    expiryDate: Date;
    employeeContactTypeId: string;
    timeStamp: any;
    totalCount: number;
    isArchived: boolean;
    employeeDetailType: string;
    employeeLicenceDetailId: string;
    licenceExpiryDate: Date;
    licenceIssuedDate: Date;
}